chore(app): notes fetching
