+++ packages/app/src/features/apollo/documents/notes.gql.tsx
@@ -1,33 +1,12 @@
 import { gql } from '@apollo/client'
 
 export const GET_ALLNOTES = gql(/* GraphQL */ `
-  query AllNotes {
-    allNotes {
+  query AllNotes($userId: String!) {
+    allNotes(userId: $userId) {
       id
+      userId
       description
       updatedAt
-    }
-  }
-`)
-
-export const GET_FINDUSER = gql(/* GraphQL */ `
-  query findUser($findUserId: String!) {
-    findUser(id: $findUserId) {
-      name
-      notes {
-        id
-        description
-        updatedAt
-      }
-    }
-  }
-`)
-
-export const POST_CREATEUSER = gql(/* GraphQL */ `
-  mutation CreateUser($data: UserCreateInput!) {
-    createUser(data: $data) {
-      name
-      email
     }
   }
 `)
